+++ CHANGE_LOG.md
@@ -6,6 +6,8 @@
 ## V2.3.0
 
 ### Major Changes
+* Add youtube-dl video download support (any site that can be downloaded by the youtube-dl project can now be downloaded
+from reddit by this downloader)
 * Massively update the failed downloads dialog
 * Add ability to export failed downloads in multiple formats
 * Add new extractor for reddituploads